+++ src/index.ts
@@ -1,4 +1,4 @@
 export * from './meta-names'
-export * from './repositories'
 export * from './interface'
 export * from './providers'
+export * from './repositories'
