+++ src/meta-names.ts
@@ -1,5 +1,4 @@
-import { ContractRepository } from './repositories/contract-repository'
-import { DomainRepository } from './repositories/domain-repository'
+import { ContractRepository, DomainRepository } from './repositories'
 import { Config, ConfigProvider, Enviroment } from './providers'
 import { MetaNamesContractRepository } from './repositories/contracts/meta-names-contract-repository'
